issue-118: Refactored code
